chore: Move UnsafeCreateAsync and UnsafeDeleteAsync to Resource
